+++ src/components/reps/date-handler.jsx
@@ -1,0 +1,4 @@
+export default {
+  shouldHandle: value => Object.prototype.toString.call(value) === '[object Date]',
+  render: value => value.toString(),
+}
