+++ src/components/reps/null-handler.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default {
+  shouldHandle: value => (value === null),
+
+  render: () => <pre>null</pre>,
+}
