+++ src/components/reps/undefined-handler.jsx
@@ -1,0 +1,6 @@
+import React from 'react'
+
+export default {
+  shouldHandle: value => (value === undefined),
+  render: () => <pre>undefined</pre>,
+}
